rename lib dir to share
