+++ src/share/defaultSettings.ts
