+++ src/share/type.d.ts
