+++ src/share/util.ts
